move outpoint to its rightful place
